+++ CHANGELOG.md
@@ -29,6 +29,9 @@
 
 ### Added
 
+- [#4084](https://github.com/ChainSafe/forest/pull/4084) Add support for the
+  `Filecoin.StateDealProviderCollateralBounds` RPC method.
+
 ### Changed
 
 ### Removed
@@ -112,6 +115,12 @@ Mandatory release that includes:
 
 - [#3955](https://github.com/ChainSafe/forest/pull/3955) Added support for the
   NV22 _Dragon_ network upgrade, together with the required state migration.
+
+### Changed
+
+- [#3976](https://github.com/ChainSafe/forest/pull/3976) `forest-wallet`
+  defaults to using a local wallet instead of the builtin Forest wallet for
+  greater security.
 
 ### Fixed